refactor: 이미지 파일 및 css 파일 구조 변경
