+++ src/components/GNB.js
@@ -1,3 +1,7 @@
+import './GNB.css';
+
+import '../statics/imgs/add-button.png';
+
 export default class GNB extends HTMLElement {
   constructor() {
     super();
