+++ src/components/Modal.js
@@ -1,3 +1,5 @@
+import './Modal.css';
+
 import { RESTAURANT_FORM_EVENTS } from './RestaurantForm';
 
 export default class Modal extends HTMLElement {
