+++ src/components/Select.js
@@ -1,3 +1,5 @@
+import './Select.css';
+
 export const SELECT_EVENTS = {
   onchange: 'selectOnChange',
 };
@@ -27,7 +29,7 @@ export default class Select extends HTMLSelectElement {
   // eslint-disable-next-line
   #handleOnChange(e) {
     const { value } = e.target;
-    
+
     this.dispatchEvent(
       new CustomEvent(SELECT_EVENTS.onchange, {
         bubbles: true,
