+++ src/index.js
@@ -1,28 +1,14 @@
-import WebController from './service/WebController';
 // components
-import GNB from './view/components/GNB';
-import Modal from './view/components/Modal';
-import Select from './view/components/Select';
-import RestaurantForm from './view/components/RestaurantForm';
-import RestaurantList from './view/components/RestaurantList';
-import RestaurantItem from './view/components/RestaurantItem';
+import App from './components/App';
+import GNB from './components/GNB';
+import Modal from './components/Modal';
+import Select from './components/Select';
+import RestaurantForm from './components/RestaurantForm';
+import RestaurantList from './components/RestaurantList';
+import RestaurantItem from './components/RestaurantItem';
 
 // styles
-import './view/styles/global.css';
-import './view/styles/GNB.css';
-import './view/styles/Modal.css';
-import './view/styles/Select.css';
-import './view/styles/RestaurantForm.css';
-import './view/styles/RestaurantList.css';
-
-// imgs
-import './view/imgs/add-button.png';
-import './view/imgs/category-korean.png';
-import './view/imgs/category-asian.png';
-import './view/imgs/category-chinese.png';
-import './view/imgs/category-japanese.png';
-import './view/imgs/category-western.png';
-import './view/imgs/category-etc.png';
+import './components/global.css';
 
 window.customElements.define('app-gnb', GNB);
 window.customElements.define('app-modal', Modal);
@@ -31,5 +17,5 @@ window.customElements.define('app-restaurant-form', RestaurantForm, { extends: '
 window.customElements.define('app-restaurant-item', RestaurantItem, { extends: 'li' });
 window.customElements.define('app-restaurant-list', RestaurantList, { extends: 'ul' });
 
-const controller = new WebController();
-controller.start();
+const app = new App();
+app.start();
