+++ src/statics/imgs/category-chinese.png
