+++ src/statics/imgs/category-japanese.png
